Move SyncPlay initialization to plugin
