+++ src/config.json
@@ -44,6 +44,7 @@
     "pdfPlayer/plugin",
     "logoScreensaver/plugin",
     "sessionPlayer/plugin",
-    "chromecastPlayer/plugin"
+    "chromecastPlayer/plugin",
+    "syncPlay/plugin"
   ]
 }
